Moving statfs calls and DriveInfo PInvokes over to use the PAL layer. The PAL layer will abstract away the knowledge of if the OS has /proc/mounts or a C API to gather the mount point information. This change also means we only need one DriveInfo.Unix file for all non-Windows platforms.
