--- src/Common/src/Interop/Linux/libc/Interop.endmntent.cs
@@ -1,14 +1,0 @@
-// Copyright (c) Microsoft. All rights reserved.
-// Licensed under the MIT license. See LICENSE file in the project root for full license information.
-
-using System;
-using System.Runtime.InteropServices;
-
-internal static partial class Interop
-{
-    internal static partial class libc
-    {
-        [DllImport(Libraries.Libc)]
-        internal static extern int endmntent(IntPtr fp);
-    }
-}
