--- src/Common/src/Interop/Linux/libc/Interop.setmntent.cs
@@ -1,17 +1,0 @@
-// Copyright (c) Microsoft. All rights reserved.
-// Licensed under the MIT license. See LICENSE file in the project root for full license information.
-
-using System;
-using System.Runtime.InteropServices;
-
-internal static partial class Interop
-{
-    internal static partial class libc
-    {
-        [DllImport(Libraries.Libc, SetLastError = true)]
-        internal static extern IntPtr setmntent(string filename, string type);
-
-        internal const string MNTTYPE_IGNORE = "ignore";
-        internal const string MNTOPT_RO = "ro";
-    }
-}
