+++ src/Native/config.h.in
@@ -5,6 +5,10 @@
 #cmakedefine01 HAVE_STAT_BIRTHTIME
 #cmakedefine01 HAVE_GNU_STRERROR_R
 #cmakedefine01 HAVE_DIRENT_NAME_LEN
+#cmakedefine01 HAVE_MNTINFO
+#cmakedefine01 HAVE_STATFS_FSTYPENAME
+#cmakedefine01 HAVE_STRCPY_S
+#cmakedefine01 HAVE_STRLCPY
 
 // Mac OS X has stat64, but it is deprecated since plain stat now
 // provides the same 64-bit aware struct when targeting OS X > 10.5
